[Refactor] 중고 및 경매상품에 공통적으로 사용되는 값에 대한 공통 패키지 생성 및 이외의 패키지 위치 변경 #144
