+++ src/main/java/com/yzgeneration/evc/domain/image/enums/ItemType.java
@@ -1,0 +1,8 @@
+package com.yzgeneration.evc.domain.image.enums;
+
+import lombok.Getter;
+
+@Getter
+public enum ItemType {
+    USEDITEM, AUCTIONITEM
+}
