+++ src/main/java/com/yzgeneration/evc/domain/item/enums/TransactionMode.java
@@ -1,0 +1,8 @@
+package com.yzgeneration.evc.domain.item.enums;
+
+import lombok.Getter;
+
+@Getter
+public enum TransactionMode {
+    SELL, BUY
+}
