+++ src/main/java/com/yzgeneration/evc/domain/item/useditem/enums/ItemStatus.java
@@ -1,0 +1,8 @@
+package com.yzgeneration.evc.domain.item.useditem.enums;
+
+import lombok.Getter;
+
+@Getter
+public enum ItemStatus {
+    ACTIVE, DELETED, BAN
+}
