--- src/main/java/com/yzgeneration/evc/domain/useditem/enums/TransactionStatue.java
@@ -1,8 +1,0 @@
-package com.yzgeneration.evc.domain.useditem.enums;
-
-import lombok.Getter;
-
-@Getter
-public enum TransactionStatue {
-    ONGOING, RESERVE, COMPLETE
-}
