--- src/main/java/com/yzgeneration/evc/domain/useditem/enums/TransactionType.java
@@ -1,8 +1,0 @@
-package com.yzgeneration.evc.domain.useditem.enums;
-
-import lombok.Getter;
-
-@Getter
-public enum TransactionType {
-    DIRECT, DELIVERY
-}
