--- src/main/java/com/yzgeneration/evc/domain/useditem/enums/UsedItemStatus.java
@@ -1,8 +1,0 @@
-package com.yzgeneration.evc.domain.useditem.enums;
-
-import lombok.Getter;
-
-@Getter
-public enum UsedItemStatus {
-    ACTIVE, DELETED, BAN
-}
