--- src/main/java/com/yzgeneration/evc/domain/useditem/model/ItemDetails.java
@@ -1,27 +1,0 @@
-package com.yzgeneration.evc.domain.useditem.model;
-
-import com.yzgeneration.evc.domain.useditem.dto.UsedItemRequest.CreateUsedItemRequest;
-import lombok.Builder;
-import lombok.Getter;
-
-@Getter
-@Builder
-public class ItemDetails {
-
-    private String title;
-
-    private String category;
-
-    private String content;
-
-    private int price;
-
-    public static ItemDetails create(CreateUsedItemRequest createUsedItemRequest) {
-        return ItemDetails.builder()
-                .title(createUsedItemRequest.getTitle())
-                .category(createUsedItemRequest.getCategory())
-                .content(createUsedItemRequest.getContent())
-                .price(createUsedItemRequest.getPrice())
-                .build();
-    }
-}
