--- src/main/java/com/yzgeneration/evc/domain/useditem/model/ItemStats.java
@@ -1,19 +1,0 @@
-package com.yzgeneration.evc.domain.useditem.model;
-
-import lombok.Builder;
-import lombok.Getter;
-
-@Getter
-@Builder
-public class ItemStats {
-
-    private int viewCount;
-
-    private int likeCount;
-
-    private int chattingCount;
-
-    public static ItemStats create() {
-        return ItemStats.builder().build();
-    }
-}
